Package data not installed without setuptools-git
This demonstrates that without setuptools-git, the package data is
not installed.

In fact, setuptools-git doesn't even need to be listed as a requirement,
it only needs to be installed in your environment. So after you have
installed, it, it will work.

It looks like having setuptools-scm has the same affect as setuptools-git.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>